% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{return_nothing}
\alias{return_nothing}
\title{Example that calls rgl}
\usage{
return_nothing()
}
\value{
Returns \code{NULL} invisibly.
}
\description{
This function has an example that calls \code{\link[rgl:open3d]{rgl::open3d()}}.
}
\examples{
if (requireNamespace("rgl", quietly = TRUE)) {
  rgl::open3d()
  rgl::close3d()
}
}
